requirements.txt
# Core dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0

# LLM Providers
groq>=0.4.0
openai>=1.3.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0

# RAG and Embeddings
sentence-transformers>=2.2.0
numpy>=1.24.0

# Utilities
pathlib>=1.0.1
